use prefix tree for initial tokenization
