+++ src/utils.hpp
@@ -80,3 +80,15 @@ public:
     std::pair<size_t, size_t> match(const std::string& orig_str, size_t curr_start);
     ~PCRE2Wrapper();
 };
+
+class Trie {
+    public:
+        Trie() = default;
+
+        void add(const std::vector<unsigned char>& str, const int value, int idx = 0);
+        int find_longest(const std::vector<unsigned char>& str, int& idx);
+
+    private:
+        std::unordered_map<unsigned char, std::unique_ptr<Trie>> m_to;
+        int m_value = -1;  // -1 for unset value
+};
